requirements.txt
aiogram==3.7.0
python-dotenv==1.0.1
groq==0.37.1

# اختيارية (مستقبلًا للويب أو لوحة تحكم)
fastapi==0.124.2
uvicorn==0.38.0
